requirements_robust.txt
# Core dependencies
pysmb>=1.2.9.1
PyMuPDF>=1.23.0
PyQt6>=6.5.0

# Security and encryption
cryptography>=41.0.0

# Additional utilities
python-dateutil>=2.8.2
ipaddress>=1.0.23

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
pylint>=2.17.0
mypy>=1.4.1